[Suraj]: Usecase 3 for updating existing contact is added.

+++ AddressBook/AddressBook/Contacts.cs
@@ -8,13 +8,13 @@ namespace AddressBook
 {
     internal class Contacts
     {
-        public string fName { get; set; }
+        public string fName { get; set; } 
         public string lName { get; set; }
         public string address { get; set; }
         public string city { get; set; }
         public string state { get; set; }
         public int zip { get; set; }
-        public int phoneNo { get; set; }
+        public double phoneNo { get; set; }
         public string email { get; set; }
     }
 }
